add toast for info and fix some things

+++ app/src/UseCase/DeleteMovieUseCase.php
@@ -2,6 +2,7 @@
 
 namespace App\UseCase;
 
+use Doctrine\DBAL\Driver\PDO\Exception;
 use Doctrine\ORM\EntityManagerInterface;
 use App\Entity\Movie;
 use Doctrine\ORM\Exception\ORMException;
@@ -18,10 +19,16 @@ class DeleteMovieUseCase
     /**
      * @throws OptimisticLockException
      * @throws ORMException
+     * @throws Exception
      */
     public function execute(int $id): void
     {
         $movie = $this->entityManager->find(Movie::class, $id);
+
+        if(!is_null($movie)) {
+            throw new Exception('Not found');
+        }
+
         $this->entityManager->remove($movie);
         $this->entityManager->flush();
     }
